+++ .gitignore
@@ -1,1 +1,2 @@
 /target
+/crates/kelpdot_macros/target
