+++ crates/kelpdot_macros/src/colors/mod.rs
@@ -1,0 +1,2 @@
+mod colors;
+pub use colors::*;
