+++ crates/kelpdot_macros/src/debug/mod.rs
@@ -1,0 +1,1 @@
+pub mod debug;
