+++ src/lib/config/loader.rs
@@ -1,7 +1,6 @@
 use kelpdot_macros::*;
 use crate::lib::{
     structs::config::KelpDotConfig,
-    terminal::{debug::debug_print},
 };
 use std::path::Path;
 /// Loads config
@@ -10,7 +9,7 @@ pub fn load_cfg(root: String) -> anyhow::Result<KelpDotConfig> {
         red!("File {}/kelp.yaml not found!", root);
         std::process::exit(1);
     }
-    debug_print(&format!("Loading config {}/kelp.yaml", root));
+    debug_print!("Loading config {}/kelp.yaml", root);
     let cfg: KelpDotConfig =
         serde_yaml::from_str(&std::fs::read_to_string(format!("{}/kelp.yaml", root))?)?;
     Ok(cfg)
