+++ src/lib/terminal/mod.rs
@@ -1,2 +1,1 @@
-pub mod debug;
 pub mod messages;
