+++ src/lib/util/os.rs
@@ -1,4 +1,4 @@
-use crate::lib::terminal::debug::debug_print;
+use kelpdot_macros::*;
 use serde::{Deserialize, Serialize};
 use std::path::Path;
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -23,7 +23,7 @@ pub fn build_os_list() -> anyhow::Result<Vec<Os>> {
     let search_paths = vec!["/etc/kelpdot/os.yaml", "/etc/kelpdot/os.yml"];
     for path in search_paths {
         if Path::new(path).exists() {
-            debug_print(&format!("Reading OS file {}", path));
+            debug_print!("Reading OS file {}", path);
             let content = std::fs::read_to_string(path)?;
             let yaml: Vec<Os> = serde_yaml::from_str(&content)?;
             for os in yaml {
@@ -31,7 +31,7 @@ pub fn build_os_list() -> anyhow::Result<Vec<Os>> {
             }
         }
     }
-    debug_print("Reading bundled OS file...");
+    debug_print!("Reading bundled OS file...");
     let bundled = include_str!("../../config/oses.yaml");
     let oses: Vec<Os> = serde_yaml::from_str(&bundled)?;
     for os in oses {
