+++ src/main.rs
@@ -7,21 +7,21 @@ use structopt::StructOpt;
 mod lib;
 use lib::{
     cli::opts::Cli,
-    terminal::{debug::debug_print, messages::not_yet_implemented},
+    terminal::{messages::not_yet_implemented},
 };
 fn main() -> anyhow::Result<()> {
     green!("KelpDot V{}", build::PKG_VERSION);
-    debug_print(&format!(
+    debug_print!(
         "Commit {} Branch {}",
         build::SHORT_COMMIT,
         build::BRANCH
-    ));
-    debug_print(&format!(
+    );
+    debug_print!(
         "Built with {}-{}",
         build::RUST_VERSION,
         build::RUST_CHANNEL
-    ));
-    debug_print(&format!("Commit by {}", build::COMMIT_AUTHOR));
+    );
+    debug_print!("Commit by {}", build::COMMIT_AUTHOR);
     green!("==============");
     // Check CLI options
     match Cli::from_args() {
